requirements.txt
streamlit==1.45.1
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pandas==2.2.3
plotly==6.1.1
google-generativeai==0.8.5
faker==37.3.0
python-dotenv==1.1.0
streamlit==1.45.1
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pandas==2.2.3
plotly==6.1.1
google-generativeai==0.8.5
groq==0.13.0
faker==37.3.0
python-dotenv==1.1.0
openpyxl==3.1.5